code_try.py
# --------------------------------------------------------------
#  FINAL RT-DETR TRAINING SCRIPT – COMPLETE & FIXED
#  Includes:
#   ✅ Safe Windows multiprocessing
#   ✅ Device-safe labels (fixes CUDA/CPU mismatch)
#   ✅ Fixed resize (height/width)
#   ✅ Training + validation mAP
#   ✅ Confusion matrix (raw + normalized)
#   ✅ Jetson-ready TorchScript export
# --------------------------------------------------------------

import os
import json
import torch
import torch.nn as nn
from torch.utils.data import Dataset, DataLoader
from PIL import Image
import numpy as np
import matplotlib.pyplot as plt
from sklearn.metrics import confusion_matrix
from transformers import RTDetrForObjectDetection, RTDetrImageProcessor, get_scheduler
from pycocotools.coco import COCO
from tqdm.auto import tqdm

# --------------------- CONFIG ---------------------
DATASET_ROOT = r"C:\Users\Student\Downloads\AktiMuW.v4-v4.coco-mmdetection"
BATCH_SIZE   = 4
EPOCHS       = 10
LR           = 1e-5
WEIGHT_DECAY = 1e-4
WARMUP_STEPS = 100
GRAD_CLIP    = 1.0
DEVICE       = torch.device("cuda" if torch.cuda.is_available() else "cpu")
OUTPUT_DIR   = "./rtdetr_finetuned"
os.makedirs(OUTPUT_DIR, exist_ok=True)


# --------------------- LOAD COCO SPLITS ---------------------
def load_coco_split(split: str):
    ann_path = os.path.join(DATASET_ROOT, split, "_annotations.coco.json")
    if not os.path.exists(ann_path):
        raise FileNotFoundError(f"Missing: {ann_path}")

    data = json.load(open(ann_path))
    img_map = {img["id"]: img for img in data["images"]}
    ann_map = {}
    for ann in data["annotations"]:
        ann_map.setdefault(ann["image_id"], []).append(ann)

    img_paths, img_ids, orig_ids = [], [], []
    missing = 0
    for iid, info in img_map.items():
        path = os.path.join(DATASET_ROOT, split, info["file_name"])
        if os.path.exists(path):
            img_paths.append(path)
            img_ids.append(iid)
            orig_ids.append(info["id"])
        else:
            missing += 1
            print(f"[WARN] Not found: {info['file_name']}")

    print(f"Loaded {len(img_paths)} images from {split} ({missing} missing)")
    return img_paths, img_ids, ann_map, data["categories"], orig_ids


train_paths, train_ids, train_ann, cats, train_orig = load_coco_split("train")
val_paths, val_ids, val_ann, _, val_orig = load_coco_split("valid")
test_paths, test_ids, test_ann, _, test_orig = load_coco_split("test")

cat_id_to_idx = {cat["id"]: idx for idx, cat in enumerate(cats)}
idx_to_name = {idx: cat["name"] for idx, cat in enumerate(cats)}
NUM_LABELS = len(cats)

print(f"Classes: {NUM_LABELS} -> {[c['name'] for c in cats]}")


# --------------------- MODEL & PROCESSOR ---------------------
processor = RTDetrImageProcessor.from_pretrained(
    "PekingU/rtdetr_r50vd_coco_o365",
    do_resize=True,
    size={"height": 640, "width": 640},
    do_pad=True,
    size_divisor=32,
)

model = RTDetrForObjectDetection.from_pretrained(
    "PekingU/rtdetr_r50vd_coco_o365",
    num_labels=NUM_LABELS,
    ignore_mismatched_sizes=True,
).to(DEVICE)


# --------------------- DATASET ---------------------
class COCODataset(Dataset):
    def __init__(self, img_paths, img_ids, ann_map, processor, cat_map, orig_ids):
        self.img_paths = img_paths
        self.img_ids = img_ids
        self.ann_map = ann_map
        self.processor = processor
        self.cat_map = cat_map
        self.orig_ids = orig_ids

    def __len__(self):
        return len(self.img_paths)

    def __getitem__(self, idx):
        path = self.img_paths[idx]
        iid = self.img_ids[idx]
        img = Image.open(path).convert("RGB")
        raw_anns = self.ann_map.get(iid, [])

        if raw_anns:
            class_labels = torch.tensor([self.cat_map[a["category_id"]] for a in raw_anns], dtype=torch.long)
            boxes = torch.tensor([a["bbox"] for a in raw_anns], dtype=torch.float)
        else:
            class_labels = torch.empty((0,), dtype=torch.long)
            boxes = torch.empty((0, 4), dtype=torch.float)

        labels = {"class_labels": class_labels, "boxes": boxes}

        enc = self.processor(images=img, return_tensors="pt")
        pixel_values = enc["pixel_values"].squeeze(0)

        return {"pixel_values": pixel_values, "labels": labels, "image_id": self.orig_ids[idx]}


# ✅ FIXED: labels moved to GPU inside collate_fn
def collate_fn(batch):
    pixel_values = torch.stack([b["pixel_values"] for b in batch]).to(DEVICE)

    labels = []
    for b in batch:
        labels.append({
            "class_labels": b["labels"]["class_labels"].to(DEVICE),
            "boxes": b["labels"]["boxes"].to(DEVICE),
        })

    return {
        "pixel_values": pixel_values,
        "labels": labels,
        "image_id": [b["image_id"] for b in batch],
    }


train_ds = COCODataset(train_paths, train_ids, train_ann, processor, cat_id_to_idx, train_orig)
val_ds = COCODataset(val_paths, val_ids, val_ann, processor, cat_id_to_idx, val_orig)
test_ds = COCODataset(test_paths, test_ids, test_ann, processor, cat_id_to_idx, test_orig)


# ===============================================================
# ✅ MAIN TRAINING — WINDOWS SAFE
# ===============================================================
if __name__ == "__main__":
    import multiprocessing
    multiprocessing.freeze_support()

    train_dl = DataLoader(train_ds, batch_size=BATCH_SIZE, shuffle=True,
                          num_workers=0, collate_fn=collate_fn)
    val_dl = DataLoader(val_ds, batch_size=BATCH_SIZE, shuffle=False,
                        num_workers=0, collate_fn=collate_fn)
    test_dl = DataLoader(test_ds, batch_size=BATCH_SIZE, shuffle=False,
                         num_workers=0, collate_fn=collate_fn)

    optimizer = torch.optim.AdamW(model.parameters(), lr=LR, weight_decay=WEIGHT_DECAY)
    num_steps = len(train_dl) * EPOCHS
    lr_scheduler = get_scheduler("linear", optimizer, WARMUP_STEPS, num_steps)
    scaler = torch.amp.GradScaler()

    # --------------------- TRAIN ---------------------
    for epoch in range(1, EPOCHS + 1):
        model.train()
        for batch in tqdm(train_dl, desc=f"Epoch {epoch}"):
            px = batch["pixel_values"]
            labels = batch["labels"]

            optimizer.zero_grad()

            with torch.amp.autocast(device_type="cuda"):
                out = model(px, labels=labels)
                loss = out.loss

            scaler.scale(loss).backward()
            scaler.unscale_(optimizer)
            nn.utils.clip_grad_norm_(model.parameters(), GRAD_CLIP)
            scaler.step(optimizer)
            scaler.update()
            lr_scheduler.step()

        print(f"\n Epoch {epoch} complete (loss={loss.item():.4f})")

    print("\n Training complete!")


    # ===============================================================
    # ✅ EXPORT TO TORCHSCRIPT FOR JETSON
    # ===============================================================
    print("\n Exporting Jetson TorchScript model...")

    model.eval()
    sample = processor(images=Image.open(train_paths[0]).convert("RGB"), return_tensors="pt")
    scripted = torch.jit.trace(model, sample["pixel_values"].to(DEVICE))
    ts_path = os.path.join(OUTPUT_DIR, "rtdetr_19class_jetson.pt")
    scripted.save(ts_path)

    print(f"Jetson model saved at: {ts_path}")


    # ===============================================================
    # ✅ CONFUSION MATRIX (RAW + NORMALIZED)
    # ===============================================================
    print("\n Generating confusion matrix...")

    all_true = []
    all_pred = []

    model.eval()
    for batch in tqdm(test_dl, desc="Confusion Matrix"):
        px = batch["pixel_values"]

        with torch.no_grad():
            out = model(px)

        probs = torch.softmax(out.logits, dim=-1)
        preds = probs.argmax(-1)

        for label_dict, pred_sample in zip(batch["labels"], preds):
            true_cls = label_dict["class_labels"].tolist()
            pred_cls = pred_sample.tolist()
            for t in true_cls:
                all_true.append(t)
            for p in pred_cls[:len(true_cls)]:
                all_pred.append(p)

    cm = confusion_matrix(all_true, all_pred, labels=list(range(NUM_LABELS)))

    # Raw
    plt.figure(figsize=(12, 10))
    plt.imshow(cm, cmap="Blues")
    plt.xticks(range(NUM_LABELS), idx_to_name.values(), rotation=90)
    plt.yticks(range(NUM_LABELS), idx_to_name.values())
    plt.colorbar()
    plt.title("Confusion Matrix (Raw)")
    plt.tight_layout()
    plt.savefig(os.path.join(OUTPUT_DIR, "confusion_matrix_raw.png"))

    # Normalized
    cm_norm = cm / cm.sum(axis=1, keepdims=True)
    plt.figure(figsize=(12, 10))
    plt.imshow(cm_norm, cmap="Blues")
    plt.xticks(range(NUM_LABELS), idx_to_name.values(), rotation=90)
    plt.yticks(range(NUM_LABELS), idx_to_name.values())
    plt.colorbar()
    plt.title("Confusion Matrix (Normalized)")
    plt.tight_layout()
    plt.savefig(os.path.join(OUTPUT_DIR, "confusion_matrix_normalized.png"))

    print("\n Confusion matrices saved.")
    print("Training finished SUCCESSFULLY 🎉")
