requirements.txt
numpy
Pillow
matplotlib
scikit-learn
tqdm
pycocotools
transformers
roboflow
torch
torchvision
